+++ internal/api/middleware/recover.go
@@ -6,10 +6,10 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-func LogErrorFuncWithRequestInfoFunc(c echo.Context, err error, stack []byte) error {
+func LogErrorFuncWithRequestInfo(c echo.Context, err error, stack []byte) error {
 	log := util.LogFromContext(c.Request().Context())
 
 	log.Error().Err(err).Bytes("stack", stack).Msg("PANIC RECOVER")
 
-	return nil
+	return err
 }